+++ src/aui-toggler/HISTORY.md
@@ -8,5 +8,6 @@ No registries yet.
 
 ## [2.5.0](https://github.com/liferay/alloy-ui/releases/tag/2.5.0)
 
+* [AUI-1388](https://issues.liferay.com/browse/AUI-1388) Make aui-toggler responsive
 * [AUI-1163](https://issues.liferay.com/browse/AUI-1163) Remove unnecessary constants
 * [AUI-1105](https://issues.liferay.com/browse/AUI-1105) Absolutely positing elements that end up in an accordian will overflow from their container when sliding